This is My first web page
In this excercise, I tried to build my first web page look alike. This is not an sophisticated page or
a professional looking page, it is just me trying my ability to position items. I used flexbox because it is
the heart of the exercise. The point is to determine my understanding of flexbox and how they work. However,
that this is a learning exercise does it not in any way make the page immune to criticism. I hope to learn
from others.

The page is a mock of an ICT training firm that provide mainly, software development training at a near free
cost.